REF: CategoricalIndex indexing tests (#33018)
