+++ pandas/tests/indexes/multi/test_formats.py
